+++ mtgjson4/provider/magic_precons.py
@@ -3,22 +3,21 @@ import json
 import logging
 import multiprocessing
 import pathlib
-import re
-from typing import Any, Dict, List
+from typing import Any, Dict, Generator, List
 
 import mtgjson4
-from mtgjson4.outputter import win_os_fix, write_deck_to_file
+import mtgjson4.util
 
 LOGGER = logging.getLogger(__name__)
 
 
-def build_and_write_decks(precon_path: str) -> None:
+def build_and_write_decks(decks_path: str) -> Generator[Dict[str, Any], None, None]:
     """
     Given the path to the precons list, this will
     compile them in MTGJSONv4 format and write out
     the decks to the "decks/" folder.
     """
-    with pathlib.Path(precon_path).open("r", encoding="utf-8") as f:
+    with pathlib.Path(decks_path).open("r", encoding="utf-8") as f:
         content = json.load(f)
 
     for deck in content:
@@ -28,6 +27,10 @@ def build_and_write_decks(precon_path: str) -> None:
             "type": deck["type"],
             "mainBoard": [],
             "sideBoard": [],
+            "meta": {
+                "version": mtgjson4.__VERSION__,
+                "date": mtgjson4.__VERSION_DATE__,
+            },
         }
 
         with multiprocessing.Pool(processes=8) as pool:
@@ -43,65 +46,34 @@ def build_and_write_decks(precon_path: str) -> None:
                 for card in cards:
                     deck_to_output["sideBoard"].append(card)
 
-        write_deck_to_file(capital_case_without_symbols(deck["name"]), deck_to_output)
         LOGGER.info("Finished set {}".format(deck["name"]))
+        yield deck_to_output
 
 
-def capital_case_without_symbols(name: str) -> str:
-    """
-    Determine the name of the output file by stripping
-    all special characters and capital casing the words.
-    :param name: Deck name (unsanitized)
-    :return: Sanitized deck name
-    """
-    word_characters_only_regex = re.compile(r"[^\w]")
-    capital_case = "".join(x for x in name.title() if not x.isspace())
-
-    return word_characters_only_regex.sub("", capital_case)
-
-
-def get_mtgjson_set_code(set_code: str) -> str:
-    """
-    Some set codes are wrong, so this will sanitize
-    the set_code passed in
-    :param set_code: Set code (unsanitized)
-    :return: Sanitized set code
-    """
-    with mtgjson4.RESOURCE_PATH.joinpath("gatherer_set_codes.json").open(
-        "r", encoding="utf-8"
-    ) as f:
-        json_dict = json.load(f)
-        for key, value in json_dict.items():
-            if set_code == value:
-                return str(key)
-
-    return win_os_fix(set_code)
-
-
-def build_single_card(precon_card: Dict[str, Any]) -> List[Dict[str, Any]]:
+def build_single_card(deck_card: Dict[str, Any]) -> List[Dict[str, Any]]:
     """
     Build the MTGJSONv4 card for each pre-con card
     in the deck.
-    :param precon_card: Card to build Precon format
+    :param deck_card: Card to build Precon format
     :return: Card(s) built from the card MTGJSONv4 format
     """
     with mtgjson4.COMPILED_OUTPUT_DIR.joinpath(
-        get_mtgjson_set_code(precon_card["set_code"].upper()) + ".json"
+        mtgjson4.util.get_mtgjson_set_code(deck_card["set_code"].upper()) + ".json"
     ).open("r") as f:
         compiled_file = json.load(f)
 
     cards = []
     for mtgjson_card in compiled_file["cards"]:
-        if "//" in precon_card["name"]:
-            if precon_card["number"][-1].isalpha():
-                precon_card["number"] = precon_card["number"][:-1]
+        if "//" in deck_card["name"]:
+            if deck_card["number"][-1].isalpha():
+                deck_card["number"] = deck_card["number"][:-1]
 
-        if mtgjson_card["number"] == precon_card["number"]:
-            mtgjson_card["count"] = precon_card["count"]
-            mtgjson_card["isFoil"] = precon_card["foil"]
+        if mtgjson_card["number"] == deck_card["number"]:
+            mtgjson_card["count"] = deck_card["count"]
+            mtgjson_card["isFoil"] = deck_card["foil"]
             cards.append(mtgjson_card)
 
     if not cards:
-        LOGGER.warning("No match for {}".format(precon_card))
+        LOGGER.warning("No match for {}".format(deck_card))
 
     return cards
